refactor: AlbumService를 Command/Query로 분리
