+++ src/main/java/com/gbsw/snapy/domain/albums/controller/AlbumController.java
@@ -6,7 +6,8 @@ import com.gbsw.snapy.domain.albums.dto.response.AlbumListResponse;
 import com.gbsw.snapy.domain.albums.dto.response.AlbumPublishResponse;
 import com.gbsw.snapy.domain.albums.dto.response.AlbumTodayResponse;
 import com.gbsw.snapy.domain.albums.dto.response.AlbumUploadResponse;
-import com.gbsw.snapy.domain.albums.service.AlbumService;
+import com.gbsw.snapy.domain.albums.service.AlbumCommandService;
+import com.gbsw.snapy.domain.albums.service.AlbumQueryService;
 import com.gbsw.snapy.global.common.ApiResponse;
 import com.gbsw.snapy.global.security.CustomUserPrincipal;
 import jakarta.validation.Valid;
@@ -22,14 +23,15 @@ import java.util.List;
 @RequestMapping("/api/albums")
 public class AlbumController {
 
-    private final AlbumService albumService;
+    private final AlbumCommandService albumCommandService;
+    private final AlbumQueryService albumQueryService;
 
     @PostMapping
     public ResponseEntity<ApiResponse<AlbumUploadResponse>> upload(
             @Valid @ModelAttribute AlbumUploadRequest request,
             @AuthenticationPrincipal CustomUserPrincipal principal
     ) {
-        AlbumUploadResponse response = albumService.upload(request, principal.getId());
+        AlbumUploadResponse response = albumCommandService.upload(request, principal.getId());
         return ResponseEntity.ok(ApiResponse.success(response));
     }
 
@@ -40,7 +42,7 @@ public class AlbumController {
             @AuthenticationPrincipal CustomUserPrincipal principal
     ) {
         Long targetUserId = (userId != null) ? userId : principal.getId();
-        List<AlbumListResponse> response = albumService.getAlbumsByMonth(targetUserId, month, principal.getId());
+        List<AlbumListResponse> response = albumQueryService.getAlbumsByMonth(targetUserId, month, principal.getId());
         return ResponseEntity.ok(ApiResponse.success(response));
     }
 
@@ -48,7 +50,7 @@ public class AlbumController {
     public ResponseEntity<ApiResponse<List<AlbumListResponse>>> getCalendarThumbnails(
             @AuthenticationPrincipal CustomUserPrincipal principal
     ) {
-        List<AlbumListResponse> response = albumService.getCalendarThumbnails(principal.getId());
+        List<AlbumListResponse> response = albumQueryService.getCalendarThumbnails(principal.getId());
         return ResponseEntity.ok(ApiResponse.success(response));
     }
 
@@ -56,7 +58,7 @@ public class AlbumController {
     public ResponseEntity<ApiResponse<AlbumTodayResponse>> getToday(
             @AuthenticationPrincipal CustomUserPrincipal principal
     ) {
-        AlbumTodayResponse response = albumService.getTodayAlbum(principal.getId());
+        AlbumTodayResponse response = albumQueryService.getTodayAlbum(principal.getId());
         return ResponseEntity.ok(ApiResponse.success(response));
     }
 
@@ -65,7 +67,7 @@ public class AlbumController {
             @PathVariable Long albumId,
             @AuthenticationPrincipal CustomUserPrincipal principal
     ) {
-        AlbumDetailResponse response = albumService.getAlbumDetail(albumId, principal.getId());
+        AlbumDetailResponse response = albumQueryService.getAlbumDetail(albumId, principal.getId());
         return ResponseEntity.ok(ApiResponse.success(response));
     }
 
@@ -74,7 +76,7 @@ public class AlbumController {
             @PathVariable Long albumId,
             @AuthenticationPrincipal CustomUserPrincipal principal
     ) {
-        AlbumPublishResponse response = albumService.publishAlbum(albumId, principal.getId());
+        AlbumPublishResponse response = albumCommandService.publishAlbum(albumId, principal.getId());
         return ResponseEntity.ok(ApiResponse.success(response));
     }
 }
