+++ src/main/java/com/gbsw/snapy/domain/albums/scheduler/AlbumScheduler.java
@@ -1,6 +1,6 @@
 package com.gbsw.snapy.domain.albums.scheduler;
 
-import com.gbsw.snapy.domain.albums.service.AlbumService;
+import com.gbsw.snapy.domain.albums.service.AlbumCommandService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -17,16 +17,16 @@ public class AlbumScheduler {
 
     private static final ZoneId KST_ZONE = ZoneId.of("Asia/Seoul");
 
-    private final AlbumService albumService;
+    private final AlbumCommandService albumCommandService;
 
     @Scheduled(cron = "0 0 0 * * *", zone = "Asia/Seoul")
     public void autoPublishDraftAlbums() {
         LocalDate today = LocalDate.now(KST_ZONE);
-        List<Long> draftAlbumIds = albumService.findDraftAlbumIdsBefore(today);
+        List<Long> draftAlbumIds = albumCommandService.findDraftAlbumIdsBefore(today);
         log.info("[AlbumScheduler] 자동 publish 대상 앨범 수: {}", draftAlbumIds.size());
         for (Long albumId : draftAlbumIds) {
             try {
-                albumService.autoPublishOne(albumId);
+                albumCommandService.autoPublishOne(albumId);
             } catch (Exception e) {
                 log.error("[AlbumScheduler] 앨범 자동 publish 실패 - albumId: {}", albumId, e);
             }
