--- src/main/java/com/gbsw/snapy/domain/albums/service/AlbumService.java
@@ -1,494 +1,0 @@
-package com.gbsw.snapy.domain.albums.service;
-
-import com.gbsw.snapy.domain.albums.dto.request.AlbumUploadRequest;
-import com.gbsw.snapy.domain.albums.dto.response.AlbumDetailResponse;
-import com.gbsw.snapy.domain.albums.dto.response.AlbumListResponse;
-import com.gbsw.snapy.domain.albums.dto.response.AlbumPublishResponse;
-import com.gbsw.snapy.domain.albums.dto.response.AlbumTodayResponse;
-import com.gbsw.snapy.domain.albums.dto.response.AlbumUploadResponse;
-import com.gbsw.snapy.domain.albums.entity.AlbumPhoto;
-import com.gbsw.snapy.domain.albums.entity.AlbumPhotoType;
-import com.gbsw.snapy.domain.albums.entity.AlbumStatus;
-import com.gbsw.snapy.domain.albums.entity.DailyAlbum;
-import com.gbsw.snapy.domain.albums.repository.AlbumPhotoRepository;
-import com.gbsw.snapy.domain.albums.repository.DailyAlbumRepository;
-import com.gbsw.snapy.domain.photos.dto.response.PhotoUploadResponse;
-import com.gbsw.snapy.domain.photos.entity.Photo;
-import com.gbsw.snapy.domain.photos.entity.PhotoType;
-import com.gbsw.snapy.domain.photos.repository.PhotoRepository;
-import com.gbsw.snapy.domain.photos.service.PhotoService;
-import com.gbsw.snapy.global.exception.CustomException;
-import com.gbsw.snapy.global.exception.ErrorCode;
-import com.gbsw.snapy.domain.stories.entity.Story;
-import com.gbsw.snapy.domain.stories.repository.StoryPhotoRepository;
-import com.gbsw.snapy.domain.stories.repository.StoryRepository;
-import com.gbsw.snapy.domain.stories.service.StoryService;
-import com.gbsw.snapy.domain.friends.repository.FriendRepository;
-import com.gbsw.snapy.domain.settings.entity.UserSetting;
-import com.gbsw.snapy.domain.settings.entity.Visibility;
-import com.gbsw.snapy.domain.settings.repository.UserSettingRepository;
-import com.gbsw.snapy.domain.notifications.event.AlbumPublishedEvent;
-import com.gbsw.snapy.domain.notifications.event.NewStoryEvent;
-import com.gbsw.snapy.infra.s3.S3Service;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionSynchronization;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.YearMonth;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.util.ArrayList;
-import java.util.EnumMap;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class AlbumService {
-
-    private final DailyAlbumRepository dailyAlbumRepository;
-    private final AlbumPhotoRepository albumPhotoRepository;
-    private final PhotoService photoService;
-    private final PhotoRepository photoRepository;
-    private final S3Service s3Service;
-    private final StoryService storyService;
-    private final StoryRepository storyRepository;
-    private final StoryPhotoRepository storyPhotoRepository;
-    private final UserSettingRepository userSettingRepository;
-    private final FriendRepository friendRepository;
-    private final ApplicationEventPublisher eventPublisher;
-    private static final ZoneId KST_ZONE = ZoneId.of("Asia/Seoul");
-
-    @Transactional
-    public AlbumUploadResponse upload(AlbumUploadRequest request, Long userId) {
-        ZonedDateTime nowKst = ZonedDateTime.now(KST_ZONE);
-        int hour = nowKst.getHour();
-
-        LocalDate today = nowKst.toLocalDate();
-        DailyAlbum album = dailyAlbumRepository.findByUserIdAndAlbumDate(userId, today)
-                .orElseGet(() -> dailyAlbumRepository.save(
-                        DailyAlbum.builder()
-                                .userId(userId)
-                                .albumDate(today)
-                                .build()
-                ));
-
-        album = dailyAlbumRepository.findByIdForUpdate(album.getId())
-                .orElseThrow(() -> new CustomException(ErrorCode.ALBUM_NOT_FOUND));
-
-        boolean publishedAlready = album.getStatus() == AlbumStatus.PUBLISHED;
-        Optional<Story> existingStory = storyRepository.findByUserIdAndAlbumId(userId, album.getId());
-
-        AlbumPhotoType resolvedType = request.getType();
-        if (!resolvedType.matches(hour)) {
-            resolvedType = findAvailableFreeSlot(album.getId(), publishedAlready, existingStory);
-        }
-
-        if (publishedAlready && existingStory.isPresent()) {
-            if (storyPhotoRepository.existsByStoryIdAndType(existingStory.get().getId(), resolvedType)) {
-                throw new CustomException(ErrorCode.DUPLICATE_ALBUM_PHOTO_TYPE);
-            }
-        } else {
-            if (albumPhotoRepository.existsByAlbumIdAndType(album.getId(), resolvedType)) {
-                throw new CustomException(ErrorCode.DUPLICATE_ALBUM_PHOTO_TYPE);
-            }
-        }
-
-        if (!publishedAlready) {
-            album.increasePhotoCount(1);
-        }
-
-        PhotoUploadResponse frontPhoto = photoService.upload(request.getFrontImage(), userId, PhotoType.FRONT);
-        PhotoUploadResponse backPhoto = photoService.upload(request.getBackImage(), userId, PhotoType.BACK);
-
-        List<String> uploadedS3Keys = List.of(frontPhoto.s3Key(), backPhoto.s3Key());
-        TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
-            @Override
-            public void afterCompletion(int status) {
-                if (status == TransactionSynchronization.STATUS_ROLLED_BACK) {
-                    for (String s3Key : uploadedS3Keys) {
-                        try {
-                            s3Service.delete(s3Key);
-                        } catch (Exception e) {
-                            log.warn("트랜잭션 롤백 후 S3 정리 실패 - key: {}", s3Key, e);
-                        }
-                    }
-                }
-            }
-        });
-
-        if (!publishedAlready) {
-            albumPhotoRepository.save(
-                    AlbumPhoto.builder()
-                            .albumId(album.getId())
-                            .photoId(frontPhoto.photoId())
-                            .type(resolvedType)
-                            .side(PhotoType.FRONT)
-                            .build()
-            );
-
-            albumPhotoRepository.save(
-                    AlbumPhoto.builder()
-                            .albumId(album.getId())
-                            .photoId(backPhoto.photoId())
-                            .type(resolvedType)
-                            .side(PhotoType.BACK)
-                            .build()
-            );
-        }
-
-        Story story;
-        boolean storyCreated = false;
-        if (existingStory.isPresent()) {
-            story = existingStory.get();
-        } else {
-            try {
-                story = storyService.createStory(userId, album.getId());
-                storyCreated = true;
-            } catch (DataIntegrityViolationException e) {
-                story = storyRepository.findByUserIdAndAlbumId(userId, album.getId())
-                        .orElseThrow(() -> e);
-            }
-        }
-        storyService.addPhotos(story.getId(), frontPhoto.photoId(), backPhoto.photoId(), resolvedType);
-
-        if (storyCreated) {
-            eventPublisher.publishEvent(new NewStoryEvent(story.getId(), userId));
-        }
-
-        return AlbumUploadResponse.from(album, resolvedType);
-    }
-
-    private AlbumPhotoType findAvailableFreeSlot(Long albumId, boolean publishedAlready,
-                                                 Optional<Story> existingStory) {
-        for (AlbumPhotoType free : List.of(AlbumPhotoType.FREE_1, AlbumPhotoType.FREE_2)) {
-            boolean taken = (publishedAlready && existingStory.isPresent())
-                    ? storyPhotoRepository.existsByStoryIdAndType(existingStory.get().getId(), free)
-                    : albumPhotoRepository.existsByAlbumIdAndType(albumId, free);
-            if (!taken) {
-                return free;
-            }
-        }
-        throw new CustomException(ErrorCode.INVALID_ALBUM_PHOTO_TIME_SLOT);
-    }
-
-    @Transactional(readOnly = true)
-    public AlbumTodayResponse getTodayAlbum(Long userId) {
-        LocalDate today = LocalDate.now(KST_ZONE);
-        DailyAlbum album = dailyAlbumRepository.findByUserIdAndAlbumDate(userId, today)
-                .orElse(null);
-        if (album == null) {
-            return null;
-        }
-
-        List<PhotoSetView> sets = loadPhotoSets(album.getId());
-        List<AlbumTodayResponse.AlbumPhotoSet> mapped = sets.stream()
-                .map(s -> new AlbumTodayResponse.AlbumPhotoSet(s.type(), s.frontImageUrl(), s.backImageUrl(), s.createdAt()))
-                .toList();
-        return AlbumTodayResponse.of(album, mapped);
-    }
-
-    @Transactional(readOnly = true)
-    public AlbumDetailResponse getAlbumDetail(Long albumId, Long userId) {
-        DailyAlbum album = dailyAlbumRepository.findById(albumId)
-                .orElseThrow(() -> new CustomException(ErrorCode.ALBUM_NOT_FOUND));
-
-        if (!album.getUserId().equals(userId)) {
-            if (album.getStatus() != AlbumStatus.PUBLISHED) {
-                throw new CustomException(ErrorCode.ACCESS_DENIED);
-            }
-
-            YearMonth albumMonth = YearMonth.from(album.getAlbumDate());
-            YearMonth currentMonth = YearMonth.now(KST_ZONE);
-            boolean isCurrentMonth = albumMonth.equals(currentMonth);
-
-            UserSetting setting = userSettingRepository.findById(album.getUserId()).orElse(null);
-
-            if (isCurrentMonth) {
-                Visibility v = (setting != null) ? setting.getFeedVisibility() : Visibility.FRIENDS_ONLY;
-                if (v == Visibility.FRIENDS_ONLY && !friendRepository.existsFriendship(userId, album.getUserId())) {
-                    throw new CustomException(ErrorCode.ACCESS_DENIED);
-                }
-            } else {
-                Visibility v = (setting != null) ? setting.getPastAlbumVisibility() : Visibility.FRIENDS_ONLY;
-                if (v == Visibility.ONLY_ME) {
-                    throw new CustomException(ErrorCode.ACCESS_DENIED);
-                }
-                if (v == Visibility.FRIENDS_ONLY && !friendRepository.existsFriendship(userId, album.getUserId())) {
-                    throw new CustomException(ErrorCode.ACCESS_DENIED);
-                }
-            }
-        }
-
-        List<PhotoSetView> sets = loadPhotoSets(album.getId());
-        List<AlbumDetailResponse.AlbumPhotoSet> mapped = sets.stream()
-                .map(s -> new AlbumDetailResponse.AlbumPhotoSet(s.type(), s.frontImageUrl(), s.backImageUrl(), s.createdAt()))
-                .toList();
-        return AlbumDetailResponse.of(album, mapped);
-    }
-
-    private List<PhotoSetView> loadPhotoSets(Long albumId) {
-        List<AlbumPhoto> albumPhotos = albumPhotoRepository.findByAlbumIdOrderByTypeAsc(albumId);
-        if (albumPhotos.isEmpty()) {
-            return List.of();
-        }
-
-        List<Long> photoIds = albumPhotos.stream().map(AlbumPhoto::getPhotoId).toList();
-        Map<Long, Photo> photoById = new HashMap<>();
-        for (Photo photo : photoRepository.findAllById(photoIds)) {
-            photoById.put(photo.getId(), photo);
-        }
-        if (photoById.size() != photoIds.size()) {
-            throw new CustomException(ErrorCode.IMAGE_NOT_FOUND);
-        }
-
-        Map<AlbumPhotoType, String> frontUrls = new EnumMap<>(AlbumPhotoType.class);
-        Map<AlbumPhotoType, String> backUrls = new EnumMap<>(AlbumPhotoType.class);
-        Map<AlbumPhotoType, LocalDateTime> createdAts = new EnumMap<>(AlbumPhotoType.class);
-        for (AlbumPhoto ap : albumPhotos) {
-            Photo photo = photoById.get(ap.getPhotoId());
-            if (ap.getSide() == PhotoType.FRONT) {
-                frontUrls.put(ap.getType(), photo.getImageUrl());
-                createdAts.put(ap.getType(), photo.getCreatedAt());
-            } else {
-                backUrls.put(ap.getType(), photo.getImageUrl());
-            }
-        }
-
-        List<PhotoSetView> sets = new ArrayList<>();
-        for (AlbumPhotoType type : AlbumPhotoType.values()) {
-            if (frontUrls.containsKey(type) || backUrls.containsKey(type)) {
-                sets.add(new PhotoSetView(type, frontUrls.get(type), backUrls.get(type),
-                        createdAts.get(type)));
-            }
-        }
-        return sets;
-    }
-
-    private record PhotoSetView(AlbumPhotoType type, String frontImageUrl, String backImageUrl,
-                                    LocalDateTime createdAt) {
-    }
-
-    @Transactional(readOnly = true)
-    public List<AlbumListResponse> getAlbumsByMonth(Long targetUserId, int month, Long requesterId) {
-        if (month < 1 || month > 12) {
-            throw new CustomException(ErrorCode.INVALID_MONTH);
-        }
-
-        boolean isOwner = targetUserId.equals(requesterId);
-        YearMonth currentMonth = YearMonth.now(KST_ZONE);
-        int year = currentMonth.getYear();
-        boolean isCurrentMonth = YearMonth.of(year, month).equals(currentMonth);
-
-        if (!isOwner) {
-            UserSetting setting = userSettingRepository.findById(targetUserId).orElse(null);
-
-            if (isCurrentMonth) {
-                Visibility feedVisibility = (setting != null) ? setting.getFeedVisibility() : Visibility.FRIENDS_ONLY;
-                if (feedVisibility == Visibility.FRIENDS_ONLY) {
-                    if (!friendRepository.existsFriendship(requesterId, targetUserId)) {
-                        throw new CustomException(ErrorCode.ACCESS_DENIED);
-                    }
-                }
-            } else {
-                Visibility albumVisibility = (setting != null) ? setting.getPastAlbumVisibility() : Visibility.FRIENDS_ONLY;
-                if (albumVisibility == Visibility.ONLY_ME) {
-                    throw new CustomException(ErrorCode.ACCESS_DENIED);
-                }
-                if (albumVisibility == Visibility.FRIENDS_ONLY) {
-                    if (!friendRepository.existsFriendship(requesterId, targetUserId)) {
-                        throw new CustomException(ErrorCode.ACCESS_DENIED);
-                    }
-                }
-            }
-        }
-        YearMonth yearMonth = YearMonth.of(year, month);
-        LocalDate start = yearMonth.atDay(1);
-        LocalDate end = yearMonth.atEndOfMonth();
-
-        List<DailyAlbum> albums = dailyAlbumRepository
-                .findByUserIdAndAlbumDateBetweenOrderByAlbumDateDesc(targetUserId, start, end);
-        if (!isOwner) {
-            albums = albums.stream()
-                    .filter(a -> a.getStatus() == AlbumStatus.PUBLISHED)
-                    .toList();
-        }
-        if (albums.isEmpty()) {
-            return List.of();
-        }
-
-        List<Long> albumIds = albums.stream().map(DailyAlbum::getId).toList();
-        List<AlbumPhoto> frontPhotos = albumPhotoRepository
-                .findByAlbumIdInAndSide(albumIds, PhotoType.FRONT);
-
-        Map<Long, AlbumPhoto> thumbnailByAlbumId = new HashMap<>();
-        for (AlbumPhoto ap : frontPhotos) {
-            AlbumPhoto current = thumbnailByAlbumId.get(ap.getAlbumId());
-            if (current == null || ap.getType().ordinal() < current.getType().ordinal()) {
-                thumbnailByAlbumId.put(ap.getAlbumId(), ap);
-            }
-        }
-
-        List<Long> photoIds = thumbnailByAlbumId.values().stream()
-                .map(AlbumPhoto::getPhotoId)
-                .toList();
-        Map<Long, Photo> photoById = new HashMap<>();
-        if (!photoIds.isEmpty()) {
-            for (Photo photo : photoRepository.findAllById(photoIds)) {
-                photoById.put(photo.getId(), photo);
-            }
-            if (photoById.size() != photoIds.size()) {
-                throw new CustomException(ErrorCode.IMAGE_NOT_FOUND);
-            }
-        }
-
-        List<AlbumListResponse> result = new ArrayList<>(albums.size());
-        for (DailyAlbum album : albums) {
-            AlbumPhoto thumbnail = thumbnailByAlbumId.get(album.getId());
-            String thumbnailUrl = null;
-            if (thumbnail != null) {
-                Photo photo = photoById.get(thumbnail.getPhotoId());
-                thumbnailUrl = photo.getImageUrl();
-            }
-            result.add(AlbumListResponse.of(album, thumbnailUrl));
-        }
-        return result;
-    }
-
-    @Transactional(readOnly = true)
-    public List<AlbumListResponse> getCalendarThumbnails(Long userId) {
-        LocalDate today = LocalDate.now(KST_ZONE);
-        LocalDate fiveMonthsAgo = today.minusMonths(5).withDayOfMonth(1);
-
-        List<DailyAlbum> albums = dailyAlbumRepository
-                .findByUserIdAndAlbumDateBetweenOrderByAlbumDateDesc(userId, fiveMonthsAgo, today);
-        if (albums.isEmpty()) {
-            return List.of();
-        }
-
-        List<Long> albumIds = albums.stream().map(DailyAlbum::getId).toList();
-        List<AlbumPhoto> frontPhotos = albumPhotoRepository
-                .findByAlbumIdInAndSide(albumIds, PhotoType.FRONT);
-
-        Map<Long, AlbumPhoto> thumbnailByAlbumId = new HashMap<>();
-        for (AlbumPhoto ap : frontPhotos) {
-            AlbumPhoto current = thumbnailByAlbumId.get(ap.getAlbumId());
-            if (current == null || ap.getType().ordinal() < current.getType().ordinal()) {
-                thumbnailByAlbumId.put(ap.getAlbumId(), ap);
-            }
-        }
-
-        List<Long> photoIds = thumbnailByAlbumId.values().stream()
-                .map(AlbumPhoto::getPhotoId)
-                .toList();
-        Map<Long, Photo> photoById = new HashMap<>();
-        if (!photoIds.isEmpty()) {
-            for (Photo photo : photoRepository.findAllById(photoIds)) {
-                photoById.put(photo.getId(), photo);
-            }
-            if (photoById.size() != photoIds.size()) {
-                throw new CustomException(ErrorCode.IMAGE_NOT_FOUND);
-            }
-        }
-
-        List<AlbumListResponse> result = new ArrayList<>(albums.size());
-        for (DailyAlbum album : albums) {
-            AlbumPhoto thumbnail = thumbnailByAlbumId.get(album.getId());
-            String thumbnailUrl = null;
-            if (thumbnail != null) {
-                Photo photo = photoById.get(thumbnail.getPhotoId());
-                thumbnailUrl = photo.getImageUrl();
-            }
-            result.add(AlbumListResponse.of(album, thumbnailUrl));
-        }
-        return result;
-    }
-
-    @Transactional
-    public AlbumPublishResponse publishAlbum(Long albumId, Long userId) {
-        DailyAlbum album = dailyAlbumRepository.findByIdForUpdate(albumId)
-                .orElseThrow(() -> new CustomException(ErrorCode.ALBUM_NOT_FOUND));
-
-        if (!album.getUserId().equals(userId)) {
-            throw new CustomException(ErrorCode.ACCESS_DENIED);
-        }
-
-        album.publish();
-
-        eventPublisher.publishEvent(new AlbumPublishedEvent(album.getId(), userId));
-
-        return AlbumPublishResponse.from(album);
-    }
-
-    @Transactional(readOnly = true)
-    public List<Long> findDraftAlbumIdsBefore(LocalDate date) {
-        return dailyAlbumRepository.findIdsByStatusAndAlbumDateBefore(AlbumStatus.DRAFT, date);
-    }
-
-    @Transactional
-    public void autoPublishOne(Long albumId) {
-        DailyAlbum album = dailyAlbumRepository.findByIdForUpdate(albumId)
-                .orElseThrow(() -> new CustomException(ErrorCode.ALBUM_NOT_FOUND));
-
-        if (album.getStatus() == AlbumStatus.PUBLISHED) {
-            return;
-        }
-
-        album.publish();
-
-        eventPublisher.publishEvent(new AlbumPublishedEvent(album.getId(), album.getUserId()));
-    }
-
-    @Transactional
-    public void deleteAlbum(Long albumId, Long userId) {
-        DailyAlbum album = dailyAlbumRepository.findById(albumId)
-                .orElseThrow(() -> new CustomException(ErrorCode.ALBUM_NOT_FOUND));
-
-        if (!album.getUserId().equals(userId)) {
-            throw new CustomException(ErrorCode.ACCESS_DENIED);
-        }
-
-        List<AlbumPhoto> albumPhotos = albumPhotoRepository.findByAlbumIdOrderByTypeAsc(album.getId());
-        List<Long> photoIds = albumPhotos.stream().map(AlbumPhoto::getPhotoId).toList();
-
-        List<Photo> photos = photoRepository.findAllById(photoIds);
-        if (photos.size() != photoIds.size()) {
-            throw new CustomException(ErrorCode.IMAGE_NOT_FOUND);
-        }
-
-        boolean hasUnauthorized = photos.stream()
-                .anyMatch(photo -> !photo.getUserId().equals(userId));
-        if (hasUnauthorized) {
-            throw new CustomException(ErrorCode.ACCESS_DENIED);
-        }
-
-        List<String> s3Keys = photos.stream().map(Photo::getS3Key).toList();
-
-        albumPhotoRepository.deleteAll(albumPhotos);
-        photoRepository.deleteAll(photos);
-        dailyAlbumRepository.delete(album);
-
-        TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
-            @Override
-            public void afterCommit() {
-                for (String s3Key : s3Keys) {
-                    try {
-                        s3Service.delete(s3Key);
-                    } catch (Exception e) {
-                        log.warn("커밋 후 S3 삭제 실패 - key: {}", s3Key, e);
-                    }
-                }
-            }
-        });
-    }
-}
